Added security control in UpdateProfileServlet.
Application will allow regular user to update only own profile (compares
posted id value with id retrieved from session attribute:
"ownProfile").This security check is performed on the server side, in
case posted id was intercepted and altered in transit by the
man-in-the-middle.
Admin will be allowed to update any user's profile.
Additionally, if the update was performed on user's own profile, the
application will update the session attribute, in order to show updated
user profile on home page, without the necessity of logging out and back
in.